test_ontology.py covers species, strain, brain-region, and mixin/round-trip behavior
