+++ tests/test_minimal/test_tools/test_ontology.py
@@ -11,25 +11,31 @@ from neuroconv.tools.ontology import (
     HBA_TERMS,
     MBA_TERMS,
     SPECIES_TERMS,
+    STRAIN_TERMS,
     BrainRegionTerm,
     SpeciesTerm,
+    StrainTerm,
     add_brain_region_external_resources,
     add_species_external_resource,
+    add_strain_external_resource,
     get_brain_region_term,
     get_species_suggestion,
     get_species_term,
+    get_strain_suggestion,
+    get_strain_term,
     validate_species,
+    validate_strain,
 )
 
 
-def _make_nwbfile(species="Mus musculus", with_subject=True) -> NWBFile:
+def _make_nwbfile(species="Mus musculus", strain=None, with_subject=True) -> NWBFile:
     nwbfile = NWBFile(
         session_description="d",
         identifier="id",
         session_start_time=datetime(2020, 1, 1, tzinfo=tzutc()),
     )
     if with_subject:
-        nwbfile.subject = Subject(subject_id="s1", species=species)
+        nwbfile.subject = Subject(subject_id="s1", species=species, strain=strain)
     return nwbfile
 
 
@@ -106,6 +112,64 @@ class TestSpeciesTerms:
     def test_validate_species_warning_points_to_bioregistry(self):
         with pytest.warns(UserWarning, match="bioregistry.io/NCBITaxon:9606"):
             validate_species("human")
+
+
+# ---------------------------------------------------------------------------
+# Strain term resolution
+# ---------------------------------------------------------------------------
+
+
+class TestStrainTerms:
+    def test_table_entries_are_self_consistent(self):
+        for canonical_name, term in STRAIN_TERMS.items():
+            assert isinstance(term, StrainTerm)
+            assert term.canonical_name == canonical_name
+            assert term.rrid.startswith("RRID:")
+
+    def test_entity_uri_is_derived_from_rrid(self):
+        term = STRAIN_TERMS["Long-Evans"]
+        assert term.rrid == "RRID:RGD_2308852"
+        assert term.entity_uri == "https://scicrunch.org/resolver/RRID:RGD_2308852"
+
+    def test_exact_canonical_name_resolves(self):
+        term = get_strain_term("Long-Evans")
+        assert term.canonical_name == "Long-Evans"
+        assert term.rrid == "RRID:RGD_2308852"
+
+    def test_informal_spelling_suggestion(self):
+        suggestion = get_strain_suggestion("black 6")
+        term, reason = suggestion
+        assert term.canonical_name == "C57BL/6J"
+        assert "informal spelling" in reason
+        assert get_strain_term("black 6").canonical_name == "C57BL/6J"
+
+    def test_informal_spelling_is_case_insensitive_and_stripped(self):
+        term, _ = get_strain_suggestion("  Long Evans  ")
+        assert term.canonical_name == "Long-Evans"
+
+    def test_typo_suggestion(self):
+        suggestion = get_strain_suggestion("Sprague Dawly")
+        term, reason = suggestion
+        assert term.canonical_name == "Sprague Dawley"
+        assert "closely matches" in reason
+
+    @pytest.mark.parametrize("strain", ["Octodon degus strain X", "", None, 42])
+    def test_unrecognized_returns_none(self, strain):
+        assert get_strain_suggestion(strain) is None
+        assert get_strain_term(strain) is None
+
+    def test_validate_strain_no_warning_for_canonical_name(self, recwarn):
+        assert validate_strain("Long-Evans") is None
+        assert len(recwarn) == 0
+
+    def test_validate_strain_warns_and_returns_term_for_informal_spelling(self):
+        with pytest.warns(UserWarning, match="C57BL/6J"):
+            term = validate_strain("black 6")
+        assert term.canonical_name == "C57BL/6J"
+
+    def test_validate_strain_warning_points_to_bioregistry(self):
+        with pytest.warns(UserWarning, match="bioregistry.io/RRID:RGD_2308852"):
+            validate_strain("long evans")
 
 
 # ---------------------------------------------------------------------------
@@ -224,6 +288,103 @@ class TestSpeciesExternalResource:
         assert add_species_external_resource(nwbfile) is True
         assert nwbfile.external_resources is herd  # extended in place, not replaced
         assert len(herd.entities[:]) == 2
+
+
+# ---------------------------------------------------------------------------
+# Strain HERD annotation
+# ---------------------------------------------------------------------------
+
+
+class TestStrainExternalResource:
+    @pytest.mark.parametrize(
+        "kwargs", [dict(with_subject=False), dict(strain=None), dict(strain="a custom in-house line")]
+    )
+    def test_noop_cases(self, kwargs):
+        nwbfile = _make_nwbfile(**kwargs)
+        assert add_strain_external_resource(nwbfile) is False
+        assert nwbfile.external_resources is None
+
+    def test_recognized_strain_is_annotated(self):
+        nwbfile = _make_nwbfile(strain="Long-Evans")
+        assert add_strain_external_resource(nwbfile) is True
+
+        dataframe = nwbfile.external_resources.to_dataframe()
+        assert dataframe["key"].tolist() == ["Long-Evans"]
+        assert dataframe["entity_id"].tolist() == ["RRID:RGD_2308852"]
+
+        objects = nwbfile.external_resources.objects.to_dataframe()
+        assert objects["object_id"].tolist() == [nwbfile.subject.object_id]
+        assert objects["relative_path"].tolist() == ["strain"]
+
+    def test_idempotent(self):
+        nwbfile = _make_nwbfile(strain="Long-Evans")
+        assert add_strain_external_resource(nwbfile) is True
+        assert add_strain_external_resource(nwbfile) is False
+        assert len(nwbfile.external_resources.entities[:]) == 1
+
+    def test_extends_existing_herd_in_place(self):
+        from hdmf.common import HERD
+        from pynwb import get_type_map
+
+        nwbfile = _make_nwbfile(strain="Long-Evans")
+        herd = HERD(type_map=get_type_map())
+        herd.add_ref(
+            container=nwbfile.subject,
+            attribute="subject_id",
+            key="s1",
+            entity_id="EXAMPLE:1",
+            entity_uri="https://example.org/1",
+        )
+        nwbfile.external_resources = herd
+
+        assert add_strain_external_resource(nwbfile) is True
+        assert nwbfile.external_resources is herd  # extended in place, not replaced
+        assert len(herd.entities[:]) == 2
+
+    def test_metadata_mapping_annotates_unrecognized_and_takes_precedence(self):
+        nwbfile = _make_nwbfile(strain="Long-Evans")
+        metadata = {
+            "Strain": {
+                # Overrides the curated result (RRID:RGD_2308852) with an explicit term.
+                "Long-Evans": {"id": "RRID:EXAMPLE:2", "uri": "https://example.org/2"},
+            }
+        }
+        assert add_strain_external_resource(nwbfile, metadata=metadata) is True
+        dataframe = nwbfile.external_resources.to_dataframe()
+        assert dataframe["entity_id"].tolist() == ["RRID:EXAMPLE:2"]
+
+    def test_maps_strain_to_multiple_ontology_terms(self):
+        nwbfile = _make_nwbfile(strain="Long-Evans")
+        metadata = {
+            "Strain": {
+                "Long-Evans": [
+                    {"id": "RRID:RGD_2308852", "uri": "https://scicrunch.org/resolver/RRID:RGD_2308852"},
+                    {"id": "RRID:EXAMPLE:3", "uri": "https://example.org/3"},
+                ]
+            }
+        }
+
+        assert add_strain_external_resource(nwbfile, metadata=metadata) is True
+        dataframe = nwbfile.external_resources.to_dataframe()
+        assert sorted(dataframe["entity_id"].tolist()) == ["RRID:EXAMPLE:3", "RRID:RGD_2308852"]
+
+    def test_metadata_mapping_without_matching_entry_falls_back_to_offline_lookup(self):
+        nwbfile = _make_nwbfile(strain="Long-Evans")
+        metadata = {"Strain": {"a different strain": {"id": "RRID:EXAMPLE:4", "uri": "https://example.org/4"}}}
+
+        assert add_strain_external_resource(nwbfile, metadata=metadata) is True
+        dataframe = nwbfile.external_resources.to_dataframe()
+        assert dataframe["entity_id"].tolist() == ["RRID:RGD_2308852"]
+
+    @pytest.mark.parametrize(
+        "bad_value",
+        ["RRID:EXAMPLE:1", {"id": "RRID:EXAMPLE:1"}, {"uri": "https://example.org/1"}, {"id": "", "uri": ""}],
+    )
+    def test_malformed_metadata_term_raises(self, bad_value):
+        nwbfile = _make_nwbfile(strain="a strain")
+        metadata = {"Strain": {"a strain": bad_value}}
+        with pytest.raises((TypeError, ValueError)):
+            add_strain_external_resource(nwbfile, metadata=metadata)
 
 
 # ---------------------------------------------------------------------------
@@ -498,7 +659,7 @@ class TestOntologyAnnotationMixin:
         assert not any(entity_id.startswith("NCBITaxon:") for entity_id in entity_ids)
         assert {"MBA:382", "MBA:385"}.issubset(set(entity_ids))  # brain-region annotation is unaffected
 
-    def test_species_and_brain_region_references_round_trip_through_file(self, tmp_path):
+    def test_species_strain_and_brain_region_references_round_trip_through_file(self, tmp_path):
         from pynwb import NWBHDF5IO
 
         interface = self._mouse_recording_interface(["CA1", "VISp"])
@@ -511,8 +672,7 @@ class TestOntologyAnnotationMixin:
             read_nwbfile = io.read()
             entity_ids = set(read_nwbfile.external_resources.to_dataframe()["entity_id"].tolist())
 
-        assert {"NCBITaxon:10090", "MBA:382", "MBA:385"}.issubset(entity_ids)
-        assert {"MBA:382", "MBA:385"}.issubset(set(entity_ids))
+        assert {"NCBITaxon:10090", "RRID:IMSR_JAX:000664", "MBA:382", "MBA:385"}.issubset(entity_ids)
 
     def test_subclass_can_override_strain_hook(self):
         from neuroconv.tools.testing.mock_interfaces import MockRecordingInterface
@@ -531,9 +691,3 @@ class TestOntologyAnnotationMixin:
         assert not any(entity_id.startswith("RRID:") for entity_id in entity_ids)
         assert "NCBITaxon:10090" in entity_ids
         assert {"MBA:382", "MBA:385"}.issubset(set(entity_ids))
-
-
-def _optical_channel():
-    from pynwb.ophys import OpticalChannel
-
-    return OpticalChannel(name="channel0", description="d", emission_lambda=500.0)
